Migrate Monday MCP to use static metadata
